+++ CHANGELOG.md
@@ -6,16 +6,20 @@ FEATURES:
 * **New Resource:** `azurerm_cosmosdb_postgresql_node_configuration` [GH-21596]
 * **New Resource:** `azurerm_cosmosdb_postgresql_role` [GH-21597]
 * **New Resource:** `azurerm_monitor_workspace` [GH-21598]
+* **New Resource:** `azurerm_network_manager_deployment` [GH-20451]
 
 ENHANCEMENTS:
 
+* provider: Resource Provider Registration now uses API Version `2022-09-01` [GH-21695]
 * provider: updating the `IsAzureStack` check to use `hashicorp/go-azure-sdk` rather than relying on the environment from `Azure/go-autorest` [GH-21697]
-* dependencies: updating to `v0.20230508.1101604` of `github.com/hashicorp/go-azure-sdk` [GH-21698]
-* dependencies: `appconfiguration` upgraded to API Version `2023-03-01` [GH-21660]
+* dependencies: updating `github.com/hashicorp/go-azure-sdk` to `v0.20230508.1171356` [GH-21707]
+* `appconfiguration`: updating to API Version `2023-03-01` [GH-21660]
 * `keyvault`: refactoring to use `hashicorp/go-azure-sdk` [GH-21621]
+* `azurerm_orbital_contact_profile` - `ip_address` is now optional [GH-21721]
 
 BUG FIXES:
 
+* provider: fixing a bug where we would invoke but not poll for the Registration State during automatic Resource Provider Registration [GH-21695]
 * `azurerm_app_configuration`: handling an API bug where when polling for `PurgeDeleted` returns a 404 rather the payload for a long-running operation [GH-21665]
 * `azurerm_eventhub_namespace` - add locks and remove unneeded WaitForState functions [GH-21656]
 * `azurerm_machine_learning_workspace` - parse `key_vault_id` insensitively [GH-21684]